refactor(control): put directional command flags into a struct

+++ lib/Server/synchro.capnp
@@ -2,10 +2,10 @@
 
 interface Synchro {
     struct Direction {
-        accelerate  @0 :Int32 = -1;
-        decelerate  @1 :Int32 = -1;
-        rotateLeft  @2 :Int32 = -1;
-        rotateRight @3 :Int32 = -1;
+        accelerate  @0 :Int8 = -1;
+        decelerate  @1 :Int8 = -1;
+        rotateLeft  @2 :Int8 = -1;
+        rotateRight @3 :Int8 = -1;
     }
 
     connect @0 () -> ();
